requirements.txt
fastapi
uvicorn[standard]
python-binance
pandas
pandas-ta
python-dotenv
pydantic
pydantic-settings
sqlalchemy
psycopg2-binary
requests
